pyproject.toml
[project]
name = "regress-stack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "crudini>=0.9.5",
    "networkx>=2.4",
    "pyroute2<0.8",
    "python-apt",
    "python-openstackclient>=7.1.4",
]

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "ruff>=0.9.5",
    "types-networkx>=3.4.2.20241227",
]

[tool.uv.sources]
python-apt = { git = "https://salsa.debian.org/apt-team/python-apt.git", rev = "2.4.y" }
